bbmerci/requirements.txt
django
pillow
crispy-tailwind
django-mathfilters
djangorestframework
psycopg2-binary
django-environ